pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-assets-extractor"
version = "0.1.0"
description = "Desktop app per estrarre font, colori, copy e asset da una pagina web pubblica."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "Pillow>=10.4.0",
  "playwright>=1.52.0",
  "PySide6>=6.8.0",
  "requests>=2.32.0",
]

[project.optional-dependencies]
build = [
  "pyinstaller>=6.13.0",
]

[project.scripts]
web-assets-extractor = "web_assets_extractor.main:main"

[tool.setuptools.packages.find]
include = ["web_assets_extractor*"]
